requirements.txt
appnope==0.1.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
cffi==1.14.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
freetype-py==2.2.0
imageio==2.9.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
notebook==6.1.5
numpy==1.19.4
packaging==20.4
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
plotly==4.12.0
prometheus-client==0.9.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pycparser==2.20
pyglet==1.5.11
Pygments==2.7.2
PyOpenGL==3.1.0
pyparsing==2.4.7
pyquaternion==0.9.9
pyrender==0.1.43
pyrsistent==0.17.3
python-dateutil==2.8.1
PyWavelets==1.1.1
pyzmq==20.0.0
qtconsole==4.7.7
QtPy==1.9.0
retrying==1.3.3
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.10.1
tornado==6.1
tqdm==4.52.0
traitlets==5.0.5
trimesh==3.8.13
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
